Refactor fish detail page and enhance data handling
- Updated FishDetailPage to utilize new hooks for fetching fish details and regulations, improving data management.
- Introduced validation functions to ensure accuracy of fishing data and regulations.
- Added skeleton loading components for better user experience during data fetching.
- Refactored components for improved readability and maintainability, including the addition of a fishing season calendar.
- Enhanced caching mechanisms for fish information and regulations to optimize performance.
