+++ src/App.tsx
@@ -28,7 +28,7 @@ import { useProfile } from "./hooks/queries";
 
 // Lazy load heavy components for better initial loading performance
 const HomePage = lazy(() => import("./pages/home"));
-const FishDetailPage = lazy(() => import("./components/fish-detail-page"));
+const FishDetailPage = lazy(() => import("./pages/fish-detail"));
 const MenuPage = lazy(() => import("./components/menu-page"));
 const SearchPage = lazy(() => import("./pages/search"));
 const WeatherPage = lazy(() => import("./components/weather-page"));
