+++ src/components/location-modal.tsx
@@ -20,6 +20,7 @@ import "leaflet/dist/leaflet.css";
 import L from "leaflet";
 import { log } from "@/lib/logging";
 import { useUpdateProfile, useUserLocation } from "@/hooks/queries";
+import { DEFAULT_LOCATION } from "@/lib/const";
 
 interface LocationData {
   latitude: number;
@@ -97,7 +98,7 @@ const MapSelection = ({
       const lat = currentLocation.latitude;
       const lng = currentLocation.longitude;
       log(
-        `Map created, centering on: ${lat}, ${lng} (${currentLocation.name})`,
+        `Map created, centering on: ${lat}, ${lng} (${currentLocation.name})`
       );
       map.setView([lat, lng], 15);
     }
@@ -113,7 +114,7 @@ const MapSelection = ({
         // Attempt to get location name via reverse geocoding
         try {
           const response = await fetch(
-            `https://nominatim.openstreetmap.org/reverse?format=json&lat=${lat}&lon=${lng}&zoom=18&addressdetails=1`,
+            `https://nominatim.openstreetmap.org/reverse?format=json&lat=${lat}&lon=${lng}&zoom=18&addressdetails=1`
           );
           const data = await response.json();
 
@@ -268,7 +269,7 @@ const LocationModal = ({
           // Attempt to get location name via reverse geocoding
           try {
             const response = await fetch(
-              `https://nominatim.openstreetmap.org/reverse?format=json&lat=${lat}&lon=${lng}&zoom=18&addressdetails=1`,
+              `https://nominatim.openstreetmap.org/reverse?format=json&lat=${lat}&lon=${lng}&zoom=18&addressdetails=1`
             );
             const data = await response.json();
             log("Reverse geocoding data:", data);
@@ -335,7 +336,7 @@ const LocationModal = ({
           enableHighAccuracy: true,
           timeout: 10000,
           maximumAge: 0,
-        },
+        }
       );
     } else {
       // If geolocation is not supported, set a default location
@@ -347,7 +348,7 @@ const LocationModal = ({
 
       log(
         "Setting default location (no geolocation support):",
-        defaultLocation,
+        defaultLocation
       );
       handleLocationUpdate(defaultLocation);
       setLoading(false);
@@ -400,7 +401,7 @@ const LocationModal = ({
         <div className="w-full rounded-md overflow-hidden h-[400px] mb-4">
           <MapSelection
             onLocationSelect={handleLocationUpdate}
-            currentLocation={currentLocation}
+            currentLocation={currentLocation || DEFAULT_LOCATION}
           />
         </div>
         <div className="">
