+++ src/components/skeletons/index.ts
@@ -1,0 +1,1 @@
+export { default as WeatherWidgetProSkeleton } from "./weather-widget-pro-skeleton";
