+++ src/hooks/queries/fish/index.ts
@@ -1,2 +1,3 @@
 export * from "./use-fish-data";
 export * from "./use-toxic-fish-data";
+export * from "./use-fish-detail";
